requirements.txt
async-generator==1.10
attrs==22.1.0
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
exceptiongroup==1.0.0rc9
google-api-core==2.11.0
google-api-python-client==2.82.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
google-cloud-bigquery==3.7.0
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
gspread==5.6.2
h11==0.14.0
idna==3.4
install==1.3.5
oauthlib==3.2.2
outcome==1.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySocks==1.7.1
pytz==2022.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
wsproto==1.2.0
pandas==1.5.2